Move randomized centralized telescope tests.
