+++ Cargo.toml
@@ -10,9 +10,9 @@ include = ["**/*.rs", "Cargo.toml", "README.md", ".gitignore"]
 
 [dependencies]
 blake2 = "0.10.6"
-rand_core = "0.6.4"
 
 [dev-dependencies]
+rand_core = "0.6.4"
 rand_chacha = "0.3.1"
 test-case = "3.3.1"
 
@@ -26,7 +26,6 @@ redundant-lifetimes = "warn"
 unit-bindings = "warn"
 unnameable-types = "warn"
 unsafe_code = "warn"
-unused_crate_dependencies = "warn"
 unused-import-braces = "warn"
 unused-lifetimes = "warn"
 unused-qualifications = "warn"
