+++ src/centralized_telescope/algorithm.rs
@@ -159,76 +159,3 @@ fn proof_hash(setup: &Setup, r: &Round) -> bool {
     let digest: Hash = hasher.finalize().into();
     sample::sample_bernoulli(&digest, setup.valid_proof_probability)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::centralized_telescope::{init, params::Params};
-    use crate::utils::test_utils::gen_items;
-    use crate::utils::types::DATA_LENGTH;
-    use rand_chacha::ChaCha20Rng;
-    use rand_core::{RngCore, SeedableRng};
-
-    #[test]
-    fn test_verify() {
-        let mut rng = ChaCha20Rng::from_seed(Default::default());
-        let nb_tests = 1_000;
-        let set_size = 1_000;
-        let params = Params {
-            soundness_param: 10.0,
-            completeness_param: 10.0,
-            set_size: 80 * set_size / 100,
-            lower_bound: 20 * set_size / 100,
-        };
-        for _t in 0..nb_tests {
-            let seed = rng.next_u32().to_be_bytes().to_vec();
-            let s_p = gen_items::<DATA_LENGTH>(&seed, set_size);
-            let setup = init::make_setup(&params);
-            let proof = prove(&setup, &s_p).unwrap();
-            assert!(verify(&setup, &proof.clone()));
-            // Checking that the proof fails if proof.search_counter is erroneous
-            let proof_t = Proof {
-                retry_counter: proof.retry_counter,
-                search_counter: proof.search_counter.wrapping_add(1),
-                element_sequence: proof.element_sequence.clone(),
-            };
-            assert!(!verify(&setup, &proof_t));
-            // Checking that the proof fails if proof.retry_counter is erroneous
-            let proof_v = Proof {
-                retry_counter: proof.retry_counter.wrapping_add(1),
-                search_counter: proof.search_counter,
-                element_sequence: proof.element_sequence.clone(),
-            };
-            assert!(!verify(&setup, &proof_v));
-            // Checking that the proof fails when no elements are included
-            let proof_item = Proof {
-                retry_counter: proof.retry_counter,
-                search_counter: proof.search_counter,
-                element_sequence: Vec::new(),
-            };
-            assert!(!verify(&setup, &proof_item));
-            // Checking that the proof fails when wrong elements are included
-            // We are trying to trigger proof_hash
-            let mut wrong_items = proof.element_sequence.clone();
-            let last_item = wrong_items.pop().unwrap();
-            let mut penultimate_item = wrong_items.pop().unwrap();
-            let proof_itembis = Proof {
-                retry_counter: proof.retry_counter,
-                search_counter: proof.search_counter,
-                element_sequence: wrong_items.clone(),
-            };
-            assert!(!verify(&setup, &proof_itembis));
-            // Checking that the proof fails when wrong elements are included
-            // We are trying to trigger round_hash
-            penultimate_item[0] = penultimate_item[0].wrapping_add(42u8);
-            wrong_items.push(penultimate_item);
-            wrong_items.push(last_item);
-            let proof_itembis = Proof {
-                retry_counter: proof.retry_counter,
-                search_counter: proof.search_counter,
-                element_sequence: wrong_items.clone(),
-            };
-            assert!(!verify(&setup, &proof_itembis));
-        }
-    }
-}
