+++ src/lib.rs
@@ -7,6 +7,3 @@ pub mod centralized_telescope;
 pub mod simple_lottery;
 
 pub mod docs;
-
-// Silence a rustc warning about unused crate.
-use rand_core as _;
