+++ src/utils/mod.rs
@@ -1,6 +1,3 @@
 pub(crate) mod sample;
 
 pub(crate) mod types;
-
-#[cfg(test)]
-pub(crate) mod test_utils;
